refactor: コア生成モジュールへの移植
