+++ voicevox_engine/tts_pipeline/__init__.py
@@ -1,12 +1,10 @@
 from ..core_wrapper import CoreWrapper, load_runtime_lib
-from .make_tts_engines import make_cores
 from .tts_engine import TTSEngine, make_tts_engines_from_cores
 from .tts_engine_base import TTSEngineBase
 
 __all__ = [
     "CoreWrapper",
     "load_runtime_lib",
-    "make_cores",
     "make_tts_engines_from_cores",
     "TTSEngine",
     "TTSEngineBase",
